Job: move some I/O helpers into _io.py
