+++ sciath/_io.py
@@ -16,3 +16,30 @@ class NamedColors:
   ENDC      = '\033[0m'
   BOLD      = '\033[1m'
   UNDERLINE = '\033[4m'
+
+# two space tab for formatted print statements
+tab = '  '
+
+# verbosity-regulated printing
+def printv(level,verbosityLevel,*vargs):
+  if level >= verbosityLevel:
+    line = ''
+    N = len(vargs)
+    for i in range(N-1):
+      line += str(vargs[i])
+      line += ' '
+    line += str(vargs[N-1])
+    print(line)
+
+def dictView(d):
+  if isinstance(d,dict):
+    string = '{'
+    for key in sorted(d):
+      value = d[key]
+      string += "'" + str(key) + "': " + str(value) + ", "
+    string = string[:-2] # remove last two characters - yes, I could have used a generator...
+    string += '}'
+    return string
+  else:
+    print('[SciATH error] dictView() requires a dictionary as input.')
+    sys.exit(1)
